refactor: extract common stdin steps to stdinExercise
