+++ lib/stdinProcessor.js
@@ -1,4 +1,3 @@
-const aliens = require('./aliens.json')
 
 function messageStdin (submissionStdin, solutionStdin, input) {
   const iv = setInterval(function () {
@@ -16,15 +15,9 @@ function messageStdin (submissionStdin, solutionStdin, input) {
 
 function stdinProcessor (exercise) {
   exercise.addProcessor(function (mode, callback) {
-    const input = []
-    for (let i = 0; i < 10; i++) {
-      const alien = aliens[Math.floor(Math.random() * aliens.length)]
-      input.push(`${alien}\n`)
-    }
-
     const solutionStdin = (mode === 'verify') ? this.solutionChild.stdin : null
 
-    messageStdin(this.submissionChild.stdin, solutionStdin, input)
+    messageStdin(this.submissionChild.stdin, solutionStdin, this.inputStdin)
 
     process.nextTick(function () {
       callback(null, true)
